+++ src/server/Movement/MovementRouteExternal.ts
@@ -5,7 +5,7 @@ const router = express.Router();
 
 /**
  * Send account movements for a specified account/year/month in an email.
- * Example: /api/external/movement/email-account-movements?account=11&year=2022&month=9
+ * Example: /api/external/movement/email-account-movements?account=11&year=2023&month=7
  */
 router.get("/email-account-movements", (req, res) => {
   const server: MovementCustom = new MovementCustom();